+++ src/Datasets/Unlabeled.php
@@ -268,7 +268,8 @@ class Unlabeled extends Dataset
     }
 
     /**
-     * Fold the dataset k - 1 times to form k equal size datasets.
+     * Fold the dataset k - 1 times to form k datasets of as equal size as
+     * possible. Any remaining samples are added to the last fold.
      *
      * @param int $k
      * @throws InvalidArgumentException
@@ -287,9 +288,11 @@ class Unlabeled extends Dataset
 
         $folds = [];
 
-        while (count($folds) < $k) {
+        while (count($folds) < $k - 1) {
             $folds[] = self::quick(array_splice($samples, 0, $n));
         }
+
+        $folds[] = self::quick($samples);
 
         return $folds;
     }