+++ CHANGELOG.md
@@ -41,6 +41,7 @@
     - Filesystem Persister now does atomic writes
     - Added cleanup() method to remove neural network residual state
     - Murmur3 new default Token Hashing Vectorizer hash function
+    - Dataset fold() now returns excess samples in last fold
     
 - 2.5.11
     - Optimize DBSCAN inference